requirements.txt
discord[voice]
riotwatcher
asyncio
requests
console-menu
colorama
mysql-connector-python
psutil